Задание к 7-му уроку

+++ src/lesson7/home/work/exercise1/Town.java
@@ -1,6 +1,6 @@
 //// ########## lesson_7 git exercise_1 ##########
-// 1. написать метод для класса улицы, который добавляет дом - добавление  неизвестного кол-ва улиц/домов
-// 1.1 написать метод для класса города, который добавляет улицу - добавление  неизвестного кол-ва улиц/домов
+// 1. написать метод для класса улицы, который добавляет дом - добавление неизвестного кол-ва улиц/домов
+// 1.1 написать метод для класса города, который добавляет улицу - добавление неизвестного кол-ва улиц/домов
 // 2. каждая улица возвращает площадь занимаемую домами
 // 2.1 каждый город возвращает площадь занимаемую улицами и площадь парков
 // 2.2 каждый город возвращает количество домов
@@ -9,17 +9,17 @@ package lesson7.home.work.exercise1;
 public class Town {
     String townName;
     int townPopulation;
-    long townArea;
+    float townArea;
     TownStreet[] townStreets;
 
-    public Town(String townName, int townPopulation, long townArea, TownStreet[] townStreets) {
+    public Town(String townName, int townPopulation, float townArea, TownStreet[] townStreets) {
         this.townName = townName;
         this.townPopulation = townPopulation;
         this.townArea = townArea;
         this.townStreets = townStreets;
     }
 
-    public void getNewTownStreets (String streetName, int streetKadastrNumber, long streetLength, long streetWidth, boolean mainStreet, TownHouse[] townHouses, TownPark[] townParks) {
+    public void getNewTownStreets (String streetName, int streetKadastrNumber, float streetLength, float streetWidth, boolean mainStreet, TownHouse[] townHouses, TownPark[] townParks) {
         if (this.townStreets == null) {
             TownStreet[] newTownStreet = new TownStreet[1];
             newTownStreet[0] = new TownStreet(streetName, streetKadastrNumber, streetLength, streetWidth, mainStreet, townHouses, townParks);
